Switch of node locs taxon spacing to double

+++ Source/mesquite/assoc/ContainedAssociates/ContainedAssociates.java
@@ -739,7 +739,7 @@ class WideTreeDrawing extends TreeDrawing  {
 		ownerModule.nodeLocsTask.calculateNodeLocs(treeDisplay,  tree, drawnRoot); //Graphics g removed as parameter May 02
 		calculateLines(tree, drawnRoot);
 		if (treeDisplay.getTaxonSpacing()<edgeWidth+2) {
-			edgeWidth= treeDisplay.getTaxonSpacing()-2;
+			edgeWidth= (int)treeDisplay.getTaxonSpacing()-2;
 			if (edgeWidth<2)
 				edgeWidth=2;
 		}
